+++ components/ui/airwayscomponents/flightCalendar.tsx
@@ -39,10 +39,10 @@ export function FlightCalendar({
             id="date"
             variant={"outline"}
             className={cn(
-              "w-full font-audimat border-0 bg-transparent hover:bg-transparent hover:text-white  rounded-none"
+              "w-full font-audimat border-0 bg-transparent hover:bg-transparent   rounded-none"
             )}
           >
-            <div className="text-4xl md:text-3xl lg:text-4xl">
+            <div className="">
               {" "}
               {/* Add margin-bottom here */}
               {!date?.from && !date?.to ? (
@@ -51,7 +51,7 @@ export function FlightCalendar({
                   <CalendarIcon size={28} className="ml-20" />
                 </div>
               ) : (
-                <div className="calendar-date text-4xl lg:text-3xl xl:text-4xl">
+                <div className=" flex ">
                   {date?.from && format(date.from, "MM/dd/yy")} -{" "}
                   {date?.to && format(date.to, "MM/dd/yy")}
                 </div>
