+++ tailwind.config.js
@@ -89,10 +89,10 @@ module.exports = {
         airlinepink: "#FF386B",
         airlineblue: "#405BFF",
         airlinetext: "#D1D3D4",
-        airlineinactive: "#939598",
+        airlinegray: "#939598",
         airlinedarkblue: "#405BFF",
         airlinelightblue: "#3DD6F5",
-        airlineBlack: "#2C2C2C",
+        airlineblack: "#212121",
         banklightblue: "#3DD6F5",
         bankdarkblue: "#405BFF",
         markettext: "#D1D3D4",
